first crack at the "morse_also" config that sends a morse cw message (callsign + grid6) after the wspr messages

+++ libraries/JTEncode/src/init_rs_int.cpp
@@ -30,7 +30,8 @@ void * JTEncode::init_rs_int(int symsize, int gfpoly, int fcr, int prim,
 
   rs = ((struct rs *)0);
   /* Check parameter ranges */
-  if(symsize < 0 || symsize > 8*sizeof(data_t)){
+  // kbn (int) for good compare
+  if(symsize < 0 || symsize > 8* (int)sizeof(data_t)){
     goto done;
   }
